basic learning algorithm functions
Here’s some examples of basic learning algorithms you can use for
linear regression. We want to train our data by minimizing an object
called the cost function. The cost is a penalty the algorithm pays for
making an incorrect classification, so naturally if we minimize this
function we will maximize the effectiveness of our learning algorithm